Ensure that functions are inlined
When optimizations are disabled, functions that only exist in header
files ended up being undefined.  The desired behavior is to force the
function to be inlined, and never compile it stand-alone.  None of the
C or C++ standards give a way to do this, but every compiler seems to
have a way.